Parse strings and comments so we can use ⟨ ⟩ in code (#2)
We're functionally equivalent to the bootstrap, but more robust
because we can handle code blocks with ⟨ and ⟩ characters in
strings (rather than using ⟦ and ⟧ and filtering them out with Perl).

Most of the code has a decent rough draft of the intended
expository prose.

I added a line_counter module and stopped using it for YAGNI
reasons.  But I'm keeping it around in the possibly foolish belief
that I will indeed re-use that code.