Use Span<lo, hi, ilk> instead of Result<>
Since error tokens aren't unexpected, we just regard Unterminated
(and future syntax error types) as just Spans to be processed in
the ordinary course of events.

Also, to disambiguate 'block' and 'section', we'll use 'block' for
a Markdown code block, and 'section' for a Knuth-like section that
can span multiple code blocks, have a name, etc.